+++ js/main.js
@@ -1,111 +1,3 @@
-// ---- DATA ---- //
-
-const PHOTOS_QUANTITY = 25;
-const MIN_LIKES = 15;
-const MAX_LIKES = 200;
-const MIN_COMMENTS = 0;
-const MAX_COMMENTS = 30;
-
-const PHOTOS_DESCRIPTIONS = [
-  '#ПоймалМомент: Когда случайность создает шедевр',
-  '#УлыбкаДня: Это фото гарантированно вызовет у вас улыбку',
-  '#НеобычныеПерспективы: Когда взгляд на мир меняется',
-  '#КошачийМарафон: Фото, которые покорят сердца всех любителей кошек',
-  '#ПриколыНаПрироде: Когда природа сама играет в шутки',
-  '#Еда-Приключение: Фото, которые заставят вас захотеть попробовать все',
-  '#СелфиСоСмыслом: Фото, которые носят глубокий смысл (или просто забавные)',
-  '#МагияСвета: Когда свет превращает обычные моменты в волшебство',
-  '#ДружбаНаВсюЖизнь: Фото, которые показывают настоящую дружбу',
-  '#ЗабавныеНеприятности: Когда все идет не по плану, но это смешно',
-  '#ВоздушныеПриключения: Фото, которые приносят ощущение свободы',
-  '#ЗабавныеМимики: Когда лица говорят сами за себя',
-  '#МагияЦветов: Фото, которые раскрывают красоту цветов',
-  '#СемейныеРадости: Когда семья – самое важное в жизни',
-  '#УловДня: Фото, которые показывают удачные моменты',
-  '#СмехДоСлез: Когда фото вызывают неудержимый смех',
-  '#АртНаПовседневности: Фото, которые превращают обычные вещи в произведения искусства',
-  '#ПриключенияРядом: Когда открываешь новые места рядом с домом',
-  '#ЗабавныеПушистики: Фото, которые покажут, что животные тоже могут быть забавными',
-  '#МоментыСчастья: Когда на фото запечатлены самые счастливые моменты',
-  '#ВзглядИзОкна: Фото, которые показывают, что красота всегда рядом',
-  '#ЗабавныеДетки: Фото, которые доказывают, что дети – наши маленькие комики',
-  '#МузыкаВДуше: Когда фото передают эмоции музыки',
-  '#ХоббиВсейЖизни: Фото, которые показывают, что хобби – это наше все',
-  '#СмешныеВидеобомбы: Когда фото взрывают ваш смех'
-];
-
-const COMMENTS__MESSAGES = [
-  'Всё отлично!',
-  'В целом всё неплохо. Но не всё.',
-  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
-  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
-  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
-  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
-];
-
-const USERS__NAMES = [
-  'Александр',
-  'Екатерина',
-  'Михаил',
-  'Анна',
-  'Дмитрий',
-  'Ольга',
-  'Иван',
-  'Наталья',
-  'Артем',
-  'Елена',
-  'Алексей',
-  'Мария',
-  'Сергей',
-  'Юлия',
-  'Владимир',
-  'Анастасия',
-  'Павел',
-  'Светлана',
-  'Николай',
-  'Ирина',
-  'Константин',
-  'Татьяна',
-  'Андрей',
-  'Оксана',
-  'Роман'
-];
-
-// ---- UTIL FUNCTIONS ---- //
-
-const getRandomInteger = (a, b) => {
-  const lower = Math.ceil(Math.min(a, b));
-  const upper = Math.floor(Math.max(a, b));
-  const result = Math.random() * (upper - lower + 1) + lower;
-  return Math.floor(result);
-};
-
-const getRandomArrayElement = (array) =>
-  array[getRandomInteger(0, array.length - 1)];
-
-// ---- MAIN FUNCTIONS ---- //
-
-const createComment = (index) => ({
-  id : index,
-  avatar: `img/avatar-${getRandomInteger(1, 6)}.svg`,
-  message : getRandomArrayElement(COMMENTS__MESSAGES),
-  name : getRandomArrayElement(USERS__NAMES)
-});
-
-const createPhoto = (index) => ({
-  id : index,
-  url : `photos/${index}.jpg`,
-  description: `${PHOTOS_DESCRIPTIONS[index - 1]}`,
-  likes: getRandomInteger(MIN_LIKES, MAX_LIKES),
-  comments : Array.from(
-    {length : getRandomInteger(MIN_COMMENTS, MAX_COMMENTS)},
-    (_, commentIndex) => createComment(commentIndex + 1)
-  ),
-});
-
-const getPictures = () =>
-  Array.from({ length: PHOTOS_QUANTITY }, (_, pictureIndex) =>
-    createPhoto(pictureIndex + 1)
-  );
+import {getPictures} from './data.js';
 
 getPictures();
